Avoid E802
Similar to 156f62c3, we are not entirely sure why this might happen, but
the cost of guarding against it is low. `:h E802` and `:h E803` bring up
the same section, so the root cause is likely similar or the same.

I decided to use a more verbose regex than necessary to make it easier
to search for this line in the code.

Fixes #1