Improve the :noh<space> fix
I had it written like this initially but didn't like it
because of some unwanted space movement behavior which
I couldn't seem to overcome with <Left>, so I tried the
version shipped in the last commit. Turns out this is
better, even if it is not perfect. Moves the cursor to
the right on each run, instead of to the second column.
Still sucks, but it is less jarring.