Open folds when jumping to a match
Folds aren't opening due to some unknown Vim quirk; perhaps there is some
conditional logic in there that prevents the default behavior from happening in
the context of a mapping. Whatever the case, adding an explicit `zv` gives us
the same effect as the default behavior.

Fixes: https://github.com/wincent/loupe/issues/9

Thanks to @adriaanzon for the report and for suggesting the fix.