scope: w:hlsearch_timer -> g:hlsearch_timer
Scoping the hlsearch_timer to a window can lead to an error if you
change windows before the timer finishes. The callback
(loupe#private#clear_hlsearch) will run in a window that has no
w:hlsearch_timer defined,

  E121: Undefined variable: w:hlsearch_timer

or it might run in a window that has an outdated w:hlsearch_timer value,
and hlsearch will never be cleared.

Since 'hlsearch' is a global option anyway, the easy fix seems to be the
correct one: make hlsearch_timer global as well.