fixes #8
The side effect of this PR, is that `/\<$SEARCH\` `/?<$SEARCH\` 
will get echoed every time `n` or `N` are pressed. 
As this is the default neovim (and maybe vim) behavior + that it isn't 
stated in https://github.com/wincent/loupe#overrides. 
I don't thinks this change in behavior matters. 
Having the search indices matches neovim defaults.