feat: add g:LoupeCaseSettingsAlways
Fixes an annoyance I've long had with `*`, `#` etc; it doesn't respect
'ignorecase' and 'smartcase' settings. It behaves as if 'ignorecase' and
'nosmartcase' were always in effect, unlike `/` and `?`.

In this commit we add a new setting, on by default, that makes all of
the search mappings behave the same, respecting the value of
'ignorecase' and 'smartcase'.

Also as noted in the comments, we take the opportunity to enhance a very
edge-casey thing: that if you have a special char like "\" in your
'iskeyword' setting, then `*` won't work when the cursor is on a `cword`
like "foo\bar". We can fix that with some escaping, seeing as we're
bypassing the whole `*`/`#` mechanism anyway.

Freshened the docs with the following (because my Haskell environment is
horrendously broken on this machine):

dist-newstyle/build/x86_64-osx/ghc-8.4.4/docvim-0.3.2.1/x/docvim/build/docvim/docvim \
  -c ~/code/loupe -v \
  ~/code/loupe/doc/loupe.txt ~/code/loupe/README.md