fix: make `:g!` work as a command
This is a little bit subtle, but I tried to explain it in the comments.
Turns out that `:g!` (and `:gl!`, `:glo!` ... `:global!`) is a command,
equivalent to `:v`, and Vim _does not_ treat the `!` as a slash
delimiter. This means that the old behavior was wrong; we would see `!`
after the `:g`, assume it was a delimiter, and append `!\v`. Instead,
what we need to do, is leave it unexpanded and wait for the user to
input a slash (or other delimiter).

Summarizing - standard usage with a slash:

- `:g/foo/d` means "delete all lines matching foo".
- `:v/foo/d` means "delete all lines not matching foo".
- `:s/foo/bar/` means "replace foo with bar".

And you can use lots of other symbols in lieu of slash, like "#":

- `:g#foo#d` means "delete all lines matching foo".
- `:v#foo#d` means "delete all lines not matching foo".
- `:s#foo#bar#` means "replace foo with bar".

But "!" is special, because sometimes it indicates a bang form of a
command and other times it can stand in for slash:

- `:g!/foo/d` means "delete all lines not matching foo"
  (ie. Vim treats this as a valid bang command).
- `:v!foo!d` is an error, because there is no `:v!` command
  (ie. Vim rejects this as an invalid bang command).
- `:s!foo!bar!` means "replace foo with bar"
  (ie. Vim does _not_ consider this a bang command, and treats "!" as a
  delimiter).
- `:g!!foo!d` means "delete all lines not matching foo"
  (ie. Vim treats this as a valid bang command _and_ treats the second
  and third "!" as delimiters).

Makes sense?

I thought so...

Updated docs with:

    bin/docvim -c ~/code/loupe ~/code/loupe/doc/loupe.txt ~/code/loupe/README.md -v

Closes: https://github.com/wincent/loupe/issues/20